requirements.txt
streamlit
langchain
langchain-openai
openai
faiss-cpu
pypdf
python-dotenv
